Make adjustments according to PR code review comments - #8307
- Move public class "UserInputDialog" to its own file
- Adjust indentation and spacing in the csproj file in the project template to improve code readability
- Use the already existing key file to sign the ProjectTemplateWizard assembly
