+++ SpecFlowStarterProjectTemplate/ProjectTemplate.csproj
@@ -8,15 +8,25 @@
     <PackageReference Include="SpecFlow" Version="3.3.30" />
     <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.3.30" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    $if$ ('$unittestframework$' == 'runner')<PackageReference Include="SpecRun.SpecFlow" Version="3.3.17" />$endif$
-    $if$ ('$unittestframework$' == 'xunit')<PackageReference Include="SpecFlow.xUnit" Version="3.3.30" />
+
+    $if$ ('$unittestframework$' == 'runner')
+    <PackageReference Include="SpecRun.SpecFlow" Version="3.3.17" />
+    $endif$
+
+    $if$ ('$unittestframework$' == 'xunit')
+    <PackageReference Include="SpecFlow.xUnit" Version="3.3.30" />
     <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />$endif$
-    $if$ ('$unittestframework$' == 'nunit')<PackageReference Include="SpecFlow.NUnit" Version="3.3.30" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
+    $endif$
+
+    $if$ ('$unittestframework$' == 'nunit')
+    <PackageReference Include="SpecFlow.NUnit" Version="3.3.30" />
     <PackageReference Include="nunit" Version="3.12.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
     $endif$
-    $if$ ('$unittestframework$' == 'mstest')<PackageReference Include="SpecFlow.MsTest" Version="3.3.30" />
+
+    $if$ ('$unittestframework$' == 'mstest')
+    <PackageReference Include="SpecFlow.MsTest" Version="3.3.30" />
     <PackageReference Include="MSTest.TestAdapter" Version="2.1.0" />
     <PackageReference Include="MSTest.TestFramework" Version="2.1.0" />
     $endif$
