+++ SpecFlowStarterProjectTemplate/SpecFlowStarterProjectTemplate.vstemplate
@@ -25,7 +25,7 @@
     </Project>
   </TemplateContent>
   <WizardExtension>
-    <Assembly>ProjectTemplateWizard, Version=2019.0.0.0, Culture=neutral, PublicKeyToken=c88ad341ebc06276</Assembly>
+    <Assembly>ProjectTemplateWizard, Version=2019.0.0.0, Culture=neutral, PublicKeyToken=0778194805d6db41</Assembly>
     <FullClassName>ProjectTemplateWizard.WizardImplementation</FullClassName>
   </WizardExtension>
 </VSTemplate>
